passgen/passgen/pyproject.toml
[tool.poetry]
name = "passgen"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.16"
pyyaml = "^6.0"
pydantic = "^1.10.9"
aiogram = "^2.25.1"
asyncpg = "^0.27.0"
greenlet = "^2.0.2"
psycopg2-binary = "^2.9.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
